Implement flexible tabu-list parameters.
tabu-parameters-f allows for setting a function that returns a list of parameters to be used as identifiers of the move. By default holds a function that takes a move and returns a list of its node-ID and vehicle-ID. By changing this function, you can change the workings of the tabu list. For instance, setting :tabu-parameters-f to #'(lambda (mv) (move-node-id mv)) will record the node-ID - this means that all moves concerning the same node is considered tabu.