Added feature to define-problem macro to accept dist-array as a list or as an array (fixes #18)
A(n asymmetric) dist-array may be passed, instead of node-coords, in which case plotting will be disabled. Dist-array must be a square list of lists or 2-dimensional array. When both node-coords and dist-array are passed, coords will be used for plotting, but dist-array will be used for calculations. When only node-coords are provided, a dist-array will be generated.

Array validation happens at the same time in same-lengthp -- additional validation rendered obsolete.